REFACTOR: move chart files to @/molecules dir
