+++ src/components/sections/index.js
@@ -1,5 +1,1 @@
-export * from './HomeBanner/HomeBanner';
-export * from './HomeFeature/HomeFeature';
-export * from './MitraList/MitraList';
-export * from './FormProgram/FormProgram';
 export * from './KonstituenType/KonstituenType';
